requirements.txt
google-cloud-storage==1.1.1
pytest==3.1.0
pytest-mock==1.6.0
pytest-cov==2.5.1
tzlocal==1.4